requirements.txt
streamlit
torch
torchvision
timm
opencv-python-headless
Pillow
requests
numpy
matplotlib
scikit-learn
ultralytics
efficientnet_pytorch
